// src/app/models/reg-ren-validate.ts
export class RegRenValidate {
  idRule?: number;
  ruleName?: string;
  documentType?: string;
  documentSection?: string;
  fieldCode?: string;
  dataType?: string;
  isRequired?: boolean;
  minValue?: number;
  maxValue?: number;
  minLength?: number;
  maxLength?: number;
  regexPattern?: string;
  allowedValues?: string;
  dependsOnField?: string;
  dependsOnValue?: string;
  errorMessage?: string;
  isActive?: boolean;
  createdAt?: string;
}